+++ drivers/esphome/driver.lua
@@ -12,6 +12,7 @@ DRIVER_FILENAMES = {
   -- #variant-filenames esphome_fan
   "esphome_climate.c4z",
   "esphome_light.c4z",
+  "esphome_camera.c4z",
   "esphome_lock.c4z",
   "esphome_switchbot.c4z",
   "esphome_yale.c4z",
@@ -42,6 +43,7 @@ local BluetoothProxyCapability = require("esphome.capabilities.bluetooth_proxy")
 
 local BinarySensorEntity = require("esphome.entities.binary_sensor")
 local ButtonEntity = require("esphome.entities.button")
+local CameraEntity = require("esphome.entities.camera")
 local ClimateEntity = require("esphome.entities.climate")
 local CoverEntity = require("esphome.entities.cover")
 local DateEntity = require("esphome.entities.date")
@@ -72,6 +74,7 @@ local bluetoothProxyCapability = BluetoothProxyCapability:new(esphome)
 local Entities = {
   [BinarySensorEntity.TYPE] = BinarySensorEntity:new(esphome),
   [ButtonEntity.TYPE] = ButtonEntity:new(esphome),
+  [CameraEntity.TYPE] = CameraEntity:new(esphome),
   [ClimateEntity.TYPE] = ClimateEntity:new(esphome),
   [CoverEntity.TYPE] = CoverEntity:new(esphome),
   [DateEntity.TYPE] = DateEntity:new(esphome),